[Refactor:System] Refactor dir setup & globals in install (#10645)
### What is the current behavior?
See issue #4011. Currently, `SUBMITTY_INSTALL_HELPER.sh`, the script
called by `submitty_install`, is a giant script that does many things to
set up the system. It is desirable to split it up into smaller scripts
so that parts can be run separately.

### What is the new behavior?
This PR moves a bunch of code that sets global variables used by the
install script to `.setup/install_submitty/get_globals.sh`, and moves
the directory setup code to
`.setup/install_submitty/setup_directories.sh`

In this PR, the code is directly cut-and-pasted to the new
`setup_directories.sh` script, which is ignored by ShellCheck. Once this
is merged, I plan to actually fix the ShellCheck issues in a new PR;
they're pretty trivial but I was advised to keep it separate from a
refactor.

### Testing
Testing was done using the Github CI, since for some reason its
environment wasn't the same as my local machine.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Barb Cutler <Barb Cutler>
Co-authored-by: Alex Lavallee <alex@lavallee.one>