[Bugfix:Developer] Fix Registration Test (#11853)
### Why is this Change Important & Necessary?
The registration test fails when within two weeks of the start of a new
term (Jan 1, and July 1) because the "recently dropped" course detects
the user is in the null section, and so it shows the "recently dropped"
course on the course list.

### What is the New Behavior?
The test has been changed to select the correct buttons to press.

It may be good to refactor the "recently dropped courses" to see if they
are self registration, and not show them, as it is kind of redundant,
but that is a slightly bigger refactor than this.