[Bugfix:System] Fix install_site Script for Vagrant VMs (#11854)
### Why is this Change Important & Necessary?
After the merging of #10645 the install_site script broke due to a
config file requirement that is supplied by install_submitty_helper but
not by install_site_helper.

### What is the New Behavior?
The submitty_install_site script should run without issue on VMs

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
There isn't testing for submitty_install_site it seems, this is an area
which should probably be looked into.

### Other information

---------

Co-authored-by: Alex Lavallee <73203142+lavalleeale@users.noreply.github.com>