[Refactor:RainbowGrades] Improve RainbowGrades handling (#11816)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
1. Refactors viewing students grades to the rainbowgrades controller,
instead of the user controller. This mimics late days being in the
latecontroller instead of the user controller
2. Added additional types to everything involved, refactoring from
`$this->getOutput()->renderOutput()` to `WebResponse`.
3. moved access control from view to controller, in addition to making
sure that instructors can see rainbowgrades when it is disabled

This removes all of the phpstan warnings about rainbowGrades
### What is the New Behavior?

<!-- Include before & after screenshots/videos if the user interface has
changed. -->
1. viewing student grades from instructor is now in rainbowgrades
controller
2. Improve rainbowgrades handling

### What steps should a reviewer take to reproduce or test the bug or
new feature?
When viewing rainbow grades for a user, the behavior should not change.
When rainbowgrades summary is disabled but instructors set up
rainbowgrades, instructors should be able to see user rainbowgrades.
Overall the page should not see that many changes.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
Issue will be made

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->