[Bugfix:System] Set Initial Gradeable Notification State (#11852)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->

The initial sample data insertion for gradeables doesn't account for the
existing notification state. Therefore, all gradeables released in the
past will have their release flag initialized as false by default,
causing the first cron job script run to release them in a large batch
for new system installations. Related to #10358.


![notification](https://github.com/user-attachments/assets/6b72ea9d-169c-4837-9844-6bed30432f0c)


### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->

The installation scripts related to inserting electronic gradeable data
for the testing courses will have the notification state set to true if
the gradeable release date is in the past, similar to how the initial
migration command handled it.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

1. Visit a course database through the CLI to reset notification states
or skip the following steps through verification via `vagrant up`
 - `$ su postgres`
 - `$ psql`
- `$ \c submitty_f25_sample` (update `f25` to suit your existing system)
- `$ UPDATE electronic_gradeable_version SET g_notification_sent =
FALSE;`
3. Run the `recreate_sample_courses` command as the root user
4. Using the initial steps in step 1, visit the testing database and run
the following command to ensure the script worked as intended.
- `SELECT * FROM electronic_gradeable_version WHERE g_notification_sent
IS TRUE;` (should be non-empty)

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

N/A

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

N/A