[Bugfix:Forum] Save chatroom in db on delete (#11858)
### Why is this Change Important & Necessary?
We want to see chatroom information in database so that professors can
check chatrooms even after they're deleted. Before deleting a chatroom
would delete it from db.

### What is the New Behavior?
Now a chatroom is kept in db, but a parameter labeled is_deleted is
switched from false to true, which prevents the chatroom from showing.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
On main, create and delete a chat. Then, go to sql toolbox and query
SELECT * FROM chatrooms. That chatroom should not exist in db anymore.
Switch to this branch, and do the same. The chatroom should still show,
but with a new column labelled "is_deleted" set to true instead of
false.

### Automated Testing & Documentation
Deleting chats will be tested once #11818 is merged

---------

Co-authored-by: Alex Lavallee <73203142+lavalleeale@users.noreply.github.com>