[Testing:Submission] Add Tutorial Course to CI (#11866)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
We would like to test some of the tutorial gradeables, and there is no
github action ci in order to do so.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Adds a github actions specifically for the tutorial gradeable. Verifies
that all the tutorial courses exist. Uses the current matrix in order to
add a new job, and then folder names are `Cypress-Autograding-{Course}`

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Look at github actions

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
Related to PR #11823