[Refactor:System] Remove pdf annotate (#11871)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
PDF annotate is a package that is maintained by the Submitty team, and
we believe that the work to maintain the package is not a good long term
plan for our team, so for now we are removing annotation support with
the hope of adding image annotation at some point in the future.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->