[Refactor:System] Move homepage to vue (#11861)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
The more components that we move to vue the easier it will be to start
having fully vue pages. The course list on the homepage is currently
pure twig.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
The course list on the homepage in now written in vue.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. View homepage on main and pr
2. Should look the same

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->