[Refactor:Notifications] Replace toasts with vue (#11857)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
As we are moving to vue this is a good early component since there are
many places on the site that make toasts manually which seem to have
drifted to cause certain toasts to have different html.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
All toasts are created in vue.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Go to main
2. Look at toasts (login/logout/submission)
3. Switch to branch
4. Look at toasts (login/logout/submission)
5. Functionality should be unchanged

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->