[Bugfix:InstructorUI] Fix missing docker image userid (#11876)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Image rights gets lost whenever an instructor uploads the same image
with a different capability. This is because whenever we check for the
current owner, an empty string is returned as the select statement
selects one string and returns the next, and there is no next string
because there can only be a single owner.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
We have a temporary value for the current row, and return that value as
part of the function

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Add a new docker image under instructor, ex `submitty/prolog:8`, with a
capability of default.
Add the same docker image under instructor `submitty/prolog:8` with a
capability of cpp.
Image owner should not change with this PR fix.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
This PR allows users to "hijack" the default images and remove them
since these images are not initially inserted into the master database.
An existing migration introduced in PR #8845
(https://github.com/Submitty/Submitty/blob/main/migration/migrator/migrations/master/20240219141515_docker_image.py)
should insert these images into the database.