[Dependency] Remove doctrine/annotations (#11880)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Before we updated to php 8, we added a package `doctrine/annotations` in
order to handle our access control and feature flags.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Now that we use php 8, we can move the annotations to attributes which
are a native php feature and we do not have to deal with updating
doctrine/annotations to a new major version.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
This should be merged after #11647 and will make #9830 outdated and
dependabot should close it.

---------

Co-authored-by: Michael Papadopoulos <micpap25@gmail.com>
Co-authored-by: William Allen <william.j.allen24@gmail.com>
Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>