[Refactor:Autograding] Invert syscall logic (#11757)
### Why is this Change Important & Necessary?
In order to autograding the Julia programming language (Issue #11621) we
need to refactor the system call to default allow rather than default
disallow.

This is continuation of work in PR #11712 (repair system call filtering)

### What is the New Behavior?
This is a refactor, so no change should be visible to the instructor or
students.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
The more_autograding_examples/c_system_call_filtering example should
work the same before & after installation of this PR. NOTE: gradeable
configuration will need to be rebuilt before & after installation.

### Automated Testing & Documentation
documentation was updated as part of the previous PR
https://github.com/Submitty/submitty.github.io/pull/680

### Other information
Once this PR is merged, we will be able to finish & merge #11718 (add
julia)

NOTE: Now that the logic is switched to default allow, it will be even
more important to stay on top of the addition of new system calls with
the release of new OS versions.
new issue: https://github.com/Submitty/Submitty/issues/11893

---------

Co-authored-by: Barb Cutler <Barb Cutler>