[Feature:Autograding] Add Julia autograding example (#11718)
### What is the new behavior?

closes issue #11621

Adds an example autograding configuration for the Julia programming
language.

The Julia programming language attempts to make a custom system call at
the start of execution. To handle this, the system call filtering used
on student code was refactored in PRs #11712 and #11757.

This autograding example is not automatically tested by either
integration tests or cypress tests. Issue #11641 requests that cypress
testing of this gradeable be added.

---------

Co-authored-by: Barb Cutler <Barb Cutler>