[Bugfix:TAGrading] Add Missing Autograding Config checks (#11894)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
When you visit /grading/status and the gradeable config has an error,
you will get a frog robot
Same thing happens for /grading/grade, which grades a specific user.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
This PR specifically checks for a missing autograding configuration
before rendering the page, ensuring that there will be no frog robot.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Break a gradeable config. Visit the page. Error. Visit the page with the
PR. It will direct you to a different page waiting for the autograding
configuration.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->