Limited Query to 5, Links Work, Added Filters, Mark Viewed
Limited Query to 5 per course (from 10). I feel like 5 is a good number because it you are in just 1 course, you will have a full panel, but no read more. This will also keep the queries pretty low, though I don't think it really matters.

Notification links now properly route to wherever the notification would route to in the course. This also marks notifications on the main page and on the course notifications page as viewed.

Added filter by unseen only, as well as mark viewed (for individual notifications that you don't want to click to) and mark all as viewed to mark all notifications as viewed.

Additionally, I adjusted the colors  in light mode to just be white with a shadow border, and it looks much better.

TODO: Refactor marking notifications as viewed so I can use it on the home page. Right now this is not feasible as it just gives you a link to its course's notification page.