[Refactor:TAGrading] Move markdownarea to vue (#11872)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Currently markdownarea uses twig and therefore when it has to be
rendered client-side for tagrading it uses twigjs which makes moving any
code that interacts with it to vue far more difficult.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Markdownarea is moved to vuejs

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->