Ran ESLint, made db limit to 10, Added hover for notifications
Ran ESlint, hence the large diff. Added a new color for notifications hover. I figured this was necessary since we will use the same on the course notifications page. Fixed the "div hell" by removing the content div and making courses and notifications distinct panels. Increased db limit to 10 per course, but today I will ask if we want to limit it to past week and limit 10 per course because we don't need 2 month old notifications hanging out on the main page.