[Bugfix:Developer] Fix Dev VM Configuration (#11910)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
As part of #9982, some things to do with the configuration has been
updated incorrect. In the PR, the dev configuration added a `y` but
since we are using pam auth, we do not even get the prompt. This leads
the dev vm to have the admin user as `y`, which then makes the
`submitty_daemon_jobs_handler` fail as some values are not set properly
in other jsons.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Gets rid of the problematic `y`

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
Tested using vagrant up on main, notice errors in
`config/submitty_admin.json`. Tested using this branch, no errors. In
addition, check the status of the service, `sudo systemctl status
submitty_daemon_jobs_handler` to see that it changed from offline to
online

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->