[Refactor:System] Move sidebar to Vue (#11859)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Slowly moving to vue is important for it to gain traction and have
examples already in the codebase.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Behavior should be the same.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Switch to main
2. View sidebar
3. Switch to PR
4. View sidebar
5. Should be unchanged

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Co-authored-by: williamschen23 <chenwill2005@gmail.com>