[Feature:InstructorUI] Team Grade Override (#10677)
## Why is this Change Important & Necessary?
fixes #10588 tested with #10697
When an instructor wants to override a team's grade, they have to do so
for each individual student.

## What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Instructors now have the option to override the grades of all students
on a team at once.

This is achieved by copying the functionality of extensions for teams;
if one student is marked as having their grade overridden, the user is
prompted with the option to mark all students on the team as overridden
with that grade.

There is now Vue confirmation popups for confirming team grade overrides
and team extensions.

![team update modal for
senger](https://github.com/user-attachments/assets/8d65a3c8-0014-4406-bdfb-accd33635afe)

## What steps should a reviewer take to reproduce or test the bug or new
feature?
tested using Rainbow Grades. I found an assignment where `Joe Student`
was on a team, and did a team override using his teammate's id. I then
generated Rainbow Grades, and observed that his grade was updated along
with his teammate's. There may be more intuitive ways to do thus, such
as using the logs. View Rainbow Grades instructions
[here](https://submitty.org/instructor/course_settings/rainbow_grades/manual_setup).

## Other information
This is not a breaking change. We can't test this in the same way as
non-team gradeables because the grade buttons on the grading page do not
display "Overridden".

---------

Co-authored-by: dorian451 <112524240+dorian451@users.noreply.github.com>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Justin Manion <jmanion32@gmail.com>