Added Media Query for Mobile, Reverted Course Title Change to Pass Cypress
The padding was too much for mobile, where all space matters. I chose 540px because that is the standard on Submitty.

I had originally changed "My Courses" to "My Active Courses" because I felt it didn't describe the page properly with the new functionality and "Courses Available for Self Registration" to just "Available for Self Registration" to avoid frequent wrapping. Because I would have to change many Cypress tests to accomodate this, I think it will be better to break it into a follow up PR.