[Feature:TAGrading] Word document to PDF conversion (#10705)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Submitted word documents need to be downloaded by grader.
Resolves #10454 

### What is the new behavior?
Word documents can now be viewed as PDF's in the browser without having
to be downloaded.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
Currently conversion takes a while on my machine (35 secs) so if it does
not seem to be converting please check `systemctl status
submitty_daemon_jobs_handler` to see the job status

---------

Co-authored-by: Alex Lavallee <alex@lavallee.one>
Co-authored-by: Alex Lavallee <73203142+lavalleeale@users.noreply.github.com>
Co-authored-by: williamschen23 <chenwill2005@gmail.com>