[Refactor:TAGrading] Refactor TA Grading Details Cookies (#11888)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Related to #11886.
TAGrading cookies have some that are gradeable specific and some that
are course specifc. This PR aims to refactor most/all of them from
gradeable-specific to course-specific. The functionality will change
slightly, including the toggles and whatnot.

In addition, the location of the code of the cookes will all be one
file, so it is easier for users to modify/change the cookies whenever
they want

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Refactor TAGrading cookies to all the course-specific.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Make sure that the functionality makes sense


https://github.com/user-attachments/assets/4d5f13de-6faf-47b5-93ca-24d4f146e671
### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
An issue will be made for testing of this refactor

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->