[Bugfix:TAGrading] Remove unused column (#11907)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
There was a regression in #11817 that added an unused column to the
grading details page.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
This PR removes the unused columns.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. View the grading details page on main and see the unused "Active
Graders" column
2. View the PR and see that it is gone

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->