[Bugfix:Autograding] Fix team autograding (#11919)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Regression came from #11578 
Current production is broken right now with team gradeables. This is
because the update portion of the team gradeable does not have a
`connection.commit()`, which the non-team gradeable has. Putting the
commit statement outside of both if and else if is going to fix the
issue.

Whenever you submit a team gradeable, you will see the `Something went
wrong will this submission` banner. When you take a closer look in the
`electronic_gradeable_data` table, the value of `completed` will always
be false. We can manually check the grades that we received in
`results/grades.txt` to see that we actually received a grade for the
assignment. As mentioned before, this is because the table is not being
updated when the job finishes, and thus the user recieves the error.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Putting the commit statement outside of both if and else if is going to
fix the issue. This ensures that insertions and updates will both be
commited.
The autoload_with=engine vs db is a semantic difference as it doesnt
impact the overall issue. SQLAlchemy takes both conection and engine,
but as engine was more widely used and recommended in sqlalchemy, we
should use this instead.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
We should do a cypress autograding test of team gradeables so that it
doesnt get broken in the future.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
check_refresh for team gradeables is broken. It will need to be fixed
before we make the autograding test, as cypress relies on the automatic
refreshs in order to detect new elements