Move business logic out of query
The query for getting the most recent notifications did not have good separation of concerns. It was making its own queries, and looping through courses. Now, all of that logic has been moved to the HomePageController. The query now returns an associative array (not an array of notification objects because we need to associate semester and course with each individual notification as well)