[Refactor:Developer] Refactor Autograding setup scripts (#11912)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
The change is necessary as every time you want to copy a file, you would
have to add it to ci, in addition to adding an rsync to the directory.
This refactor aims to combine the two elements by putting it directly
into `install_submitty.sh`, which both ci and developers run

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Whenever we are in the ci, we should copy the files. whenever we are a
developer vm, we should rsync the files.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->